+++ src/net/server/handlers/login/ViewAllCharSelectedWithPicHandler.java
@@ -20,8 +20,8 @@ public class ViewAllCharSelectedWithPicHandler extends AbstractMaplePacketHandle
 
     String pic = slea.readMapleAsciiString();
     int charId = slea.readInt();
-    int world = slea.readInt();//world
-                int channel = Randomizer.rand(0, Server.getInstance().getWorld(world).getChannels().size());
+    byte world = (byte) slea.readInt();//world
+                byte channel = (byte) Randomizer.rand(0, Server.getInstance().getWorld(world).getChannels().size());
 		String macs = slea.readMapleAsciiString();
 		c.updateMacs(macs);
